backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
beautifulsoup4==4.14.3
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
dataclasses-json==0.6.7
distro==1.9.0
fastapi==0.128.0
filelock==3.20.3
frozenlist==1.8.0
fsspec==2026.1.0
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.2.6
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.7
langchain-text-splitters==1.1.0
langgraph==1.0.6
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.6
langgraph-sdk==0.3.3
langsmith==0.6.4
lxml==6.0.2
MarkupSafe==3.0.3
marshmallow==3.26.2
mpmath==1.3.0
multidict==6.7.0
mypy_extensions==1.1.0
networkx==3.6.1
numpy==2.4.1
oauthlib==3.3.1
openai==2.15.0
orjson==3.11.5
ormsgpack==1.12.1
packaging==25.0
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
pydantic-settings==2.12.0
PySocks==1.7.1
python-dotenv==1.2.1
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
safetensors==0.7.0
setuptools==80.9.0
sniffio==1.3.1
snscrape==0.7.0.20230622
soupsieve==2.8.1
SQLAlchemy==2.0.45
starlette==0.50.0
sympy==1.14.0
tenacity==9.1.2
tokenizers==0.22.2
torch==2.9.1
tqdm==4.67.1
transformers==4.57.6
tweepy==4.16.0
typing_extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.2
urllib3==2.6.3
uuid_utils==0.13.0
uvicorn==0.40.0
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
